Fixed literal declarations in lalrpop

+++ src/parser/expression_parser.rs
@@ -12567,7 +12567,9 @@ pub fn __action21<
     (_, _, _): (usize, &'input str, usize),
 ) -> Expression
 {
-    Expression::SignedBitVector( SignedBitVectorData { size: 8, value: i64::from_str(i).unwrap() } )
+    Expression::SignedBitVector( SignedBitVectorData {
+        size: 8, value: i8::from_str(i).unwrap() as i64
+    } )
 }
 
 pub fn __action22<
@@ -12579,7 +12581,9 @@ pub fn __action22<
     (_, _, _): (usize, &'input str, usize),
 ) -> Expression
 {
-    Expression::SignedBitVector( SignedBitVectorData { size: 16, value: i64::from_str(i).unwrap() } )
+    Expression::SignedBitVector( SignedBitVectorData {
+        size: 16, value: i16::from_str(i).unwrap() as i64
+    } )
 }
 
 pub fn __action23<
@@ -12591,7 +12595,9 @@ pub fn __action23<
     (_, _, _): (usize, &'input str, usize),
 ) -> Expression
 {
-    Expression::SignedBitVector( SignedBitVectorData { size: 32, value: i64::from_str(i).unwrap() } )
+    Expression::SignedBitVector( SignedBitVectorData {
+        size: 32, value: i32::from_str(i).unwrap() as i64
+    } )
 }
 
 pub fn __action24<
@@ -12603,7 +12609,9 @@ pub fn __action24<
     (_, _, _): (usize, &'input str, usize),
 ) -> Expression
 {
-    Expression::SignedBitVector( SignedBitVectorData { size: 64, value: i64::from_str(i).unwrap() } )
+    Expression::SignedBitVector( SignedBitVectorData {
+        size: 64, value: i64::from_str(i).unwrap() as i64
+    } )
 }
 
 pub fn __action25<
@@ -12615,7 +12623,9 @@ pub fn __action25<
     (_, _, _): (usize, &'input str, usize),
 ) -> Expression
 {
-    Expression::SignedBitVector( SignedBitVectorData { size: 8, value: i64::from_str(i).unwrap() } )
+    Expression::SignedBitVector( SignedBitVectorData {
+        size: 8, value: i8::from_str(i).unwrap() as i64
+    } )
 }
 
 pub fn __action26<
@@ -12627,7 +12637,9 @@ pub fn __action26<
     (_, _, _): (usize, &'input str, usize),
 ) -> Expression
 {
-    Expression::SignedBitVector( SignedBitVectorData { size: 16, value: i64::from_str(i).unwrap() } )
+    Expression::SignedBitVector( SignedBitVectorData {
+        size: 16, value: i16::from_str(i).unwrap() as i64
+    } )
 }
 
 pub fn __action27<
@@ -12639,7 +12651,9 @@ pub fn __action27<
     (_, _, _): (usize, &'input str, usize),
 ) -> Expression
 {
-    Expression::SignedBitVector( SignedBitVectorData { size: 32, value: i64::from_str(i).unwrap() } )
+    Expression::SignedBitVector( SignedBitVectorData {
+        size: 32, value: i32::from_str(i).unwrap() as i64
+    } )
 }
 
 pub fn __action28<
@@ -12651,7 +12665,9 @@ pub fn __action28<
     (_, _, _): (usize, &'input str, usize),
 ) -> Expression
 {
-    Expression::SignedBitVector( SignedBitVectorData { size: 64, value: i64::from_str(i).unwrap() } )
+    Expression::SignedBitVector( SignedBitVectorData {
+        size: 64, value: i64::from_str(i).unwrap() as i64
+    } )
 }
 
 pub fn __action29<
@@ -12663,7 +12679,9 @@ pub fn __action29<
     (_, _, _): (usize, &'input str, usize),
 ) -> Expression
 {
-    Expression::UnsignedBitVector( UnsignedBitVectorData { size: 8, value: u64::from_str(i).unwrap() } )
+    Expression::UnsignedBitVector( UnsignedBitVectorData {
+        size: 8, value: u8::from_str(i).unwrap() as u64
+    } )
 }
 
 pub fn __action30<
@@ -12675,7 +12693,9 @@ pub fn __action30<
     (_, _, _): (usize, &'input str, usize),
 ) -> Expression
 {
-    Expression::UnsignedBitVector( UnsignedBitVectorData { size: 16, value: u64::from_str(i).unwrap() } )
+    Expression::UnsignedBitVector( UnsignedBitVectorData {
+        size: 16, value: u16::from_str(i).unwrap() as u64
+    } )
 }
 
 pub fn __action31<
@@ -12687,7 +12707,9 @@ pub fn __action31<
     (_, _, _): (usize, &'input str, usize),
 ) -> Expression
 {
-    Expression::UnsignedBitVector( UnsignedBitVectorData { size: 32, value: u64::from_str(i).unwrap() } )
+    Expression::UnsignedBitVector( UnsignedBitVectorData {
+        size: 32, value: u32::from_str(i).unwrap() as u64
+    } )
 }
 
 pub fn __action32<
@@ -12699,7 +12721,9 @@ pub fn __action32<
     (_, _, _): (usize, &'input str, usize),
 ) -> Expression
 {
-    Expression::UnsignedBitVector( UnsignedBitVectorData { size: 64, value: u64::from_str(i).unwrap() } )
+    Expression::UnsignedBitVector( UnsignedBitVectorData {
+        size: 64, value: u64::from_str(i).unwrap() as u64
+    } )
 }
 
 pub fn __action33<
@@ -12711,7 +12735,9 @@ pub fn __action33<
     (_, t, _): (usize, String, usize),
 ) -> Expression
 {
-    Expression::VariableMapping( VariableMappingData { name: n, var_type: t } )
+    Expression::VariableMapping( VariableMappingData {
+        name: n, var_type: t
+    } )
 }
 
 pub fn __action34<
